pyproject.toml
[project]
name = "rag-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.2.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.0",
    "google-genai>=1.46.0",
    "ipykernel>=7.0.1",
    "jq>=1.10.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]
